+++ lib/online_migrations/verbose_sql_logs.rb
@@ -13,7 +13,13 @@ module OnlineMigrations
         stdout_logger.level = @activerecord_logger_was.level
         stdout_logger = ActiveSupport::TaggedLogging.new(stdout_logger)
 
-        combined_logger = stdout_logger.extend(ActiveSupport::Logger.broadcast(@activerecord_logger_was))
+        combined_logger =
+          # Broadcasting logs API was changed in https://github.com/rails/rails/pull/48615.
+          if Utils.ar_version >= 7.1
+            ActiveSupport::BroadcastLogger.new(stdout_logger, @activerecord_logger_was)
+          else
+            stdout_logger.extend(ActiveSupport::Logger.broadcast(@activerecord_logger_was))
+          end
 
         ActiveRecord::Base.logger = combined_logger
         set_verbose_query_logs(false)